+++ lib/chess/move/barriers.ex
@@ -5,8 +5,14 @@ defmodule Chess.Move.Barriers do
 
   defmacro __using__(_opts) do
     quote do
-      defp check_barriers_on_route(squares, move_from, route, distance) do
-        check_squares_for_barrier(squares, calc_from_index(coordinates(move_from)), calc_direction(route), distance, 1)
+      defp do_check_barriers_on_route(squares, move_from, [route, distance]) do
+        check_squares_for_barrier(
+          squares,
+          move_from |> coordinates() |> calc_from_index(),
+          calc_direction(route),
+          distance,
+          1
+        )
       end
 
       defp coordinates(move_from), do: String.split(move_from, "", trim: true)
@@ -30,19 +36,15 @@ defmodule Chess.Move.Barriers do
       defp check_squares_for_barrier(squares, [move_from_x_index, move_from_y_index], [x_direction, y_direction], distance, step) when distance > step do
         move_to_x_index = Kernel.trunc(move_from_x_index + step * x_direction)
         move_to_y_index = Kernel.trunc(move_from_y_index + step * y_direction)
+        square = "#{Enum.at(@x_fields, move_to_x_index)}#{Enum.at(@y_fields, move_to_y_index)}"
 
-        "#{Enum.at(@x_fields, move_to_x_index)}#{Enum.at(@y_fields, move_to_y_index)}"
-        |> check_square(squares)
-
-        check_squares_for_barrier(squares, [move_from_x_index, move_from_y_index], [x_direction, y_direction], distance, step + 1)
+        case squares[:"#{square}"] != nil do
+          true -> {:error, "There is barrier at square #{square}"}
+          false -> check_squares_for_barrier(squares, [move_from_x_index, move_from_y_index], [x_direction, y_direction], distance, step + 1)
+        end
       end
 
-      defp check_squares_for_barrier(_, _, _, _, _) do
-      end
-
-      defp check_square(new_square, squares) do
-        if squares[:"#{new_square}"] != nil, do: raise "There is barrier at square #{new_square}"
-      end
+      defp check_squares_for_barrier(_, _, _, _, _), do: {:ok}
     end
   end
 end